+++ src/components/RouterApp.tsx
@@ -1,8 +1,13 @@
+import { ErrorBoundary } from '@/components/ErrorBoundary';
 import { useAuth } from '@/lib/AuthProvider';
 import { router } from '@/routes';
 import { RouterProvider } from '@tanstack/react-router';
 
 export function RouterApp() {
   const auth = useAuth();
-  return <RouterProvider router={router} context={{ auth }} />;
+  return (
+    <ErrorBoundary>
+      <RouterProvider router={router} context={{ auth }} />
+    </ErrorBoundary>
+  );
 }